+++ NEWS.md
@@ -5,6 +5,7 @@ emmeans 1.4.2.0009xxx
 
   * Added `"bcnPower"` option to `make.tran()` (per `car::bcnPower()`)
   * Scoping correction for `emmtrends()` (#153)
+  * Allow passing `...` to hook functions (need exposed by #154)
 
 
 emmeans 1.4.2